smx_analytics/requirements.txt
amqp==2.5.2
anyjson==0.3.3
astroid==2.2.5
Babel==2.7.0
billiard==3.6.1.0
blessings==1.7
cached-property==1.5.1
celery==4.4.0rc4
certifi==2019.9.11
chardet==3.0.4
colorama==0.4.1
curtsies==0.3.0
dj-database-url==0.5.0
Django==2.1.2
django-appconf==1.0.3
django-celery==3.3.1
django-celery-beat==1.5.0
django-celery-results==1.1.2
django-cors-headers==3.2.0
django-extensions==2.2.3
django-filter==2.2.0
django-haystack==2.8.1
django-pandas==0.6.1
django-timezone-field==3.1
django-url-filter==0.3.13
djangorestframework==3.10.3
dnspython==1.16.0
enum-compat==0.0.3
flower==0.9.3
gunicorn==20.0.4
idna==2.8
importlib-metadata==0.23
isort==4.3.18
kombu==4.6.6
lazy-object-proxy==1.3.1
mccabe==0.6.1
monotonic==1.5
more-itertools==7.2.0
numpy==1.17.3
pandas==0.25.3
pep8==1.7.1
Pillow==7.0.0
psycopg2==2.8.4
Pygments==2.5.2
pylint==2.3.1
python-crontab==2.4.0
python-dateutil==2.8.0
python-decouple==3.1
pytz==2019.1
PyYAML==5.1.2
redis==3.3.11
requests==2.22.0
six==1.12.0
tornado==5.1.1
typed-ast==1.4.0
typing==3.7.4.1
urllib3==1.25.6
vine==1.3.0
wcwidth==0.1.7
websockets==8.1
Whoosh==2.7.4
wrapt==1.11.1
zipp==0.6.0
